+++ src/main/java/com/cta4j/bus/direction/internal/impl/DirectionsApiImpl.java
@@ -8,7 +8,7 @@ import com.cta4j.bus.direction.internal.wire.CtaDirection;
 import com.cta4j.bus.internal.wire.CtaError;
 import com.cta4j.bus.internal.wire.CtaResponse;
 import com.cta4j.exception.Cta4jException;
-import com.cta4j.internal.http.HttpClient;
+import com.cta4j.common.internal.http.HttpClient;
 import org.apache.hc.core5.net.URIBuilder;
 import org.jetbrains.annotations.ApiStatus;
 import org.jspecify.annotations.NullMarked;