+++ src/main/java/com/cta4j/bus/locale/internal/impl/LocalesApiImpl.java
@@ -10,7 +10,7 @@ import com.cta4j.bus.locale.internal.wire.CtaLocale;
 import com.cta4j.bus.locale.internal.mapper.SupportedLocaleMapper;
 import com.cta4j.bus.locale.model.SupportedLocale;
 import com.cta4j.exception.Cta4jException;
-import com.cta4j.internal.http.HttpClient;
+import com.cta4j.common.internal.http.HttpClient;
 import org.apache.hc.core5.net.URIBuilder;
 import org.jetbrains.annotations.ApiStatus;
 import org.jspecify.annotations.NullMarked;