+++ src/main/java/com/cta4j/train/common/internal/wire/CtaResponse.java
@@ -1,4 +1,4 @@
-package com.cta4j.train.internal.wire;
+package com.cta4j.train.common.internal.wire;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import org.jetbrains.annotations.ApiStatus;