+++ src/main/java/com/cta4j/train/follow/internal/wire/CtaFollowResponse.java
@@ -1,6 +1,6 @@
 package com.cta4j.train.follow.internal.wire;
 
-import com.cta4j.train.internal.wire.CtaArrival;
+import com.cta4j.train.common.internal.wire.CtaArrival;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import org.jetbrains.annotations.ApiStatus;
 import org.jspecify.annotations.NullMarked;