+++ CHANGELOG.md
@@ -170,7 +170,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `fromExternal` method from `StationArrival` enum.
 - `fromExternal` method from `TrainCoordinates` class.
 - `fromExternal` method from `UpcomingTrainArrival` enum.
-- `com.cta4j.internal.http` export from `module-info.java`.
+- `com.cta4j.common.internal.http` export from `module-info.java`.
 
 ## 1.0.0 - 2025-10-10